// js/stat-evaluation-player/src/generated/WallAerialStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type WallAerialStats = { count: number, high_confidence_count: number, is_last_wall_aerial: boolean, last_wall_aerial_time: number | null, last_wall_aerial_frame: number | null, time_since_last_wall_aerial: number | null, frames_since_last_wall_aerial: number | null, last_confidence: number | null, best_confidence: number, cumulative_confidence: number, cumulative_setup_duration: number, cumulative_takeoff_to_touch_time: number, cumulative_touch_height: number, };